add image sizes to inline photo

+++ kotlin-telegram-framework/src/main/kotlin/io/cutebot/telegram/client/model/photo/TgPhotoSize.kt
@@ -14,7 +14,7 @@ class TgPhotoSize (
         val height: Int,
 
         @field: JsonProperty("file_size")
-        val fileSize: Int?,
+        val fileSize: Long?,
 
         fileId: String,
